Switch from Bazel to Webpack 4
While Bazel is great, it's also a bit unnecessarily convoluted for this particular project. Webpack does the job, arguably a lot better.

This also moves around resources for a more meaningful structure:
- remove public/ directory, as the build system doesn't actually copy from it (each copy is explicit, so we might as well put manifest.json and images in the root directory)
- move Popup.tsx/css to components/Popup.tsx/css
- rename index.css/tsx to popup.css/tsx to reflect the fact that it's the extension's popup, not a general index page
- rename index.html to popup.html and move it next to popup.tsx
- update CircleCI config accordingly
