+++ .gitignore
@@ -7,7 +7,7 @@
 /coverage
 
 # production
-/build
+/dist
 
 # misc
 .DS_Store
@@ -15,9 +15,6 @@
 .env.development.local
 .env.test.local
 .env.production.local
-
-# Bazel
-bazel-*
 
 npm-debug.log*
 yarn-debug.log*
