--- BUILD.bazel
@@ -1,48 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-load("@bazel_javascript//:defs.bzl", "npm_packages")
-
-exports_files([
-  "tsconfig.json",
-])
-
-npm_packages(
-  name = "packages",
-  package_json = ":package.json",
-  yarn_lock = ":yarn.lock",
-)
-
-genrule(
-  name = "extension-unpacked",
-  srcs = [
-    "//public:images",
-    "//public:manifest.json",
-    "//src:popup-prod_bundle",
-    "//src:background-prod_bundle",
-  ],
-  outs = ["extension-dir"],
-  cmd = """
-  # Create the directory.
-  mkdir $@
-  # Copy manifest.json.
-  cp $(location //public:manifest.json) $@/
-  # Copy images.
-  mkdir $@/images
-  cp -r $(location //public:images) $@/images/
-  # Copy background script.
-  cp -r $(location //src:background-prod_bundle)/* $@/
-  # Copy popup bundle.
-  cp -r $(location //src:popup-prod_bundle)/* $@/
-  """,
-)
-
-genrule(
-  name = "extension-packed",
-  srcs = [":extension-unpacked"],
-  outs = ["extension.zip"],
-  cmd = """
-  origin=$$(pwd)
-  cd $(location //:extension-unpacked)
-  zip -r $$origin/$@ *
-  """,
-)
