--- public/BUILD.bazel
@@ -1,11 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-exports_files([
-  "index.html",
-  "manifest.json",
-])
-
-filegroup(
-  name = "images",
-  srcs = glob(["images/*.png"])
-)
