--- src/BUILD.bazel
@@ -1,90 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-load("@bazel_javascript//:defs.bzl", "web_bundle", "ts_library")
-
-web_bundle(
-  name = "background-dev",
-  lib = ":background",
-  entry = "background.js",
-  output = "background.js",
-  mode = "development",
-  split_chunks = 0,
-)
-
-web_bundle(
-  name = "background-prod",
-  lib = ":background",
-  entry = "background.js",
-  output = "background.js",
-  mode = "production",
-  split_chunks = 0,
-)
-
-ts_library(
-  name = "background",
-  srcs = [
-    "background.ts",
-  ],
-  deps = [
-    ":auth",
-    "//src/github",
-    "//:packages"
-  ],
-  tsconfig = "//:tsconfig.json",
-)
-
-web_bundle(
-  name = "popup-dev",
-  lib = ":popup",
-  entry = "index.js",
-  output = "popup.js",
-  mode = "development",
-  html_template = "//public:index.html",
-)
-
-web_bundle(
-  name = "popup-prod",
-  lib = ":popup",
-  entry = "index.js",
-  output = "popup.js",
-  mode = "production",
-  html_template = "//public:index.html",
-)
-
-ts_library(
-  name = "popup",
-  srcs = [
-    "index.tsx",
-    "index.css",
-  ],
-  deps = [
-    ":components_popup",
-    "//:packages",
-  ],
-  tsconfig = "//:tsconfig.json",
-)
-
-ts_library(
-  name = "components_popup",
-  srcs = [
-    "Popup.tsx",
-    "Popup.css",
-  ],
-  deps = [
-    ":auth",
-    "//src/github",
-    "//:packages",
-  ],
-  tsconfig = "//:tsconfig.json",
-)
-
-ts_library(
-  name = "auth",
-  srcs = [
-    "auth.ts",
-  ],
-  deps = [
-    "//:packages",
-  ],
-  tsconfig = "//:tsconfig.json",
-)
