--- src/github/BUILD.bazel
@@ -1,12 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-load("@bazel_javascript//:defs.bzl", "ts_library")
-
-ts_library(
-  name = "github",
-  srcs = glob(["*.ts"]),
-  deps = [
-    "//:packages"
-  ],
-  tsconfig = "//:tsconfig.json",
-)
