--- src/index.tsx
@@ -1,6 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import "./index.css";
-import Popup from "./Popup";
-
-ReactDOM.render(<Popup />, document.getElementById("root"));
